Rename and align with other reactive examples
