+++ apps/cookbook/src/app/examples/form-field-example/examples/textarea/reactive-forms.ts
@@ -1,26 +1,31 @@
-import { Component, Input } from '@angular/core';
+import { Component } from '@angular/core';
 import { FormBuilder, FormGroup, ReactiveFormsModule } from '@angular/forms';
-import { FormFieldModule, InputComponent } from '@kirbydesign/designsystem/form-field';
+import { FormFieldModule, TextareaComponent } from '@kirbydesign/designsystem/form-field';
 import { ButtonComponent } from '@kirbydesign/designsystem/button';
-import { InputSize } from '@kirbydesign/designsystem';
+import { CheckboxComponent } from '@kirbydesign/designsystem';
 import { ReactiveFormStateComponent } from '../../../reactive-form-state/reactive-form-state.component';
 import { ExampleConfigurationWrapperComponent } from '../../../example-configuration-wrapper/example-configuration-wrapper.component';
 
 const config = {
-  selector: 'cookbook-form-field-input-counter-form-example',
+  selector: 'cookbook-form-field-textarea-counter-form-example',
   template: `<form [formGroup]="form">
   <kirby-form-field>
-    <input 
-      kirby-input 
-      [size]="size" 
+    <textarea 
+      kirby-textarea 
       placeholder="Enter your message (max 140 chars)" 
       #message 
       maxlength="140"
       formControlName="message"
-    />
+    ></textarea>
     <kirby-input-counter [listenTo]="message"></kirby-input-counter>
   </kirby-form-field>
-  
+</form>
+<cookbook-example-configuration-wrapper>
+  <kirby-checkbox
+    [checked]="true"
+    (checkedChange)="toggleEnabled()"
+    text="Form field enabled"
+  ></kirby-checkbox>
   <button 
     kirby-button 
     attentionLevel="3" 
@@ -29,8 +34,6 @@ const config = {
   >
     Reset
   </button>
-</form>
-<cookbook-example-configuration-wrapper>
   <cookbook-reactive-form-state [form]="form"></cookbook-reactive-form-state>
 </cookbook-example-configuration-wrapper>`,
 };
@@ -40,22 +43,32 @@ const config = {
   template: config.template,
   imports: [
     FormFieldModule,
-    InputComponent,
+    TextareaComponent,
     ReactiveFormsModule,
     ButtonComponent,
+    CheckboxComponent,
     ReactiveFormStateComponent,
     ExampleConfigurationWrapperComponent,
   ],
 })
-export class FormFieldInputCounterFormExampleComponent {
+export class FormFieldTextareaCounterFormExampleComponent {
   template: string = config.template;
   form: FormGroup;
-  @Input() size: InputSize;
+  isEnabled = true;
 
   constructor(private fb: FormBuilder) {
     this.form = this.fb.group({
       message: [''],
     });
+  }
+
+  toggleEnabled() {
+    this.isEnabled = !this.isEnabled;
+    if (this.isEnabled) {
+      this.form.enable();
+    } else {
+      this.form.disable();
+    }
   }
 
   resetForm() {
