+++ apps/cookbook/src/app/examples/form-field-example/form-field-example.component.ts
@@ -11,7 +11,7 @@ import { FormFieldInputLabelExampleComponent } from './examples/input/label';
 import { FormFieldInputLabelMessageExampleComponent } from './examples/input/label-message';
 import { FormFieldInputAffixExampleComponent } from './examples/input/affix';
 import { FormFieldInputCounterExampleComponent } from './examples/input/counter';
-import { FormFieldInputCounterFormExampleComponent } from './examples/input/counter-form';
+import { FormFieldInputCounterFormExampleComponent } from './examples/input/reactive-forms';
 import { FormFieldInputNumericExampleComponent } from './examples/input/numeric';
 import { FormFieldInputDateExampleComponent } from './examples/input/date';
 import { FormFieldInputDateNativeExampleComponent } from './examples/input/date-native';
@@ -22,7 +22,7 @@ import { FormFieldFocusExampleComponent } from './examples/input/focus';
 import { FormFieldTextareaDefaultExampleComponent } from './examples/textarea/default';
 import { FormFieldTextareaLabelExampleComponent } from './examples/textarea/label';
 import { FormFieldTextareaCounterExampleComponent } from './examples/textarea/counter';
-import { FormFieldTextareaCounterFormExampleComponent } from './examples/textarea/counter-form';
+import { FormFieldTextareaCounterFormExampleComponent } from './examples/textarea/reactive-forms';
 
 @Component({
   selector: 'cookbook-form-field-example',
