+++ apps/cookbook/src/app/showcase/form-field-showcase/form-field-showcase.component.ts
@@ -8,7 +8,7 @@ import { FormFieldInputLabelExampleComponent } from '../../examples/form-field-e
 import { FormFieldInputLabelMessageExampleComponent } from '../../examples/form-field-example/examples/input/label-message';
 import { FormFieldInputAffixExampleComponent } from '../../examples/form-field-example/examples/input/affix';
 import { FormFieldInputCounterExampleComponent } from '../../examples/form-field-example/examples/input/counter';
-import { FormFieldInputCounterFormExampleComponent } from '../../examples/form-field-example/examples/input/counter-form';
+import { FormFieldInputCounterFormExampleComponent } from '../../examples/form-field-example/examples/input/reactive-forms';
 import { FormFieldInputNumericExampleComponent } from '../../examples/form-field-example/examples/input/numeric';
 import { FormFieldInputDecimalMaskExampleComponent } from '../../examples/form-field-example/examples/input/decimal-mask';
 import { FormFieldInputDateExampleComponent } from '../../examples/form-field-example/examples/input/date';
@@ -20,7 +20,7 @@ import { FormFieldFocusExampleComponent } from '../../examples/form-field-exampl
 import { FormFieldTextareaDefaultExampleComponent } from '../../examples/form-field-example/examples/textarea/default';
 import { FormFieldTextareaLabelExampleComponent } from '../../examples/form-field-example/examples/textarea/label';
 import { FormFieldTextareaCounterExampleComponent } from '../../examples/form-field-example/examples/textarea/counter';
-import { FormFieldTextareaCounterFormExampleComponent } from '../../examples/form-field-example/examples/textarea/counter-form';
+import { FormFieldTextareaCounterFormExampleComponent } from '../../examples/form-field-example/examples/textarea/reactive-forms';
 import { ApiDescriptionPropertiesComponent } from '../../shared/api-description/api-description-properties/api-description-properties.component';
 import { ApiDescriptionMethodsComponent } from '../../shared/api-description/api-description-methods/api-description-methods.component';
 import { ApiDescriptionEventsComponent } from '../../shared/api-description/api-description-events/api-description-events.component';
